Split exactly once and at least once and add a corresponding property
